Do not initiate policy engine incase of --config-only flag

+++ pkg/cli/run.go
@@ -187,7 +187,7 @@ func (s *ScanOptions) Run() error {
 	}
 
 	// executor output
-	results, err := executor.Execute()
+	results, err := executor.Execute(s.configOnly)
 	if err != nil {
 		return err
 	}
@@ -204,7 +204,7 @@ func (s *ScanOptions) Run() error {
 		return err
 	}
 
-	if results.Violations.ViolationStore.Summary.ViolatedPolicies != 0 && flag.Lookup("test.v") == nil {
+	if !s.configOnly && results.Violations.ViolationStore.Summary.ViolatedPolicies != 0 && flag.Lookup("test.v") == nil {
 		os.RemoveAll(tempDir)
 		os.Exit(3)
 	}
@@ -228,6 +228,13 @@ func (s *ScanOptions) downloadRemoteRepository(tempDir string) error {
 }
 
 func (s ScanOptions) writeResults(results runtime.Output) error {
+
+	outputWriter := NewOutputWriter(s.UseColors)
+
+	if s.configOnly {
+		return writer.Write(s.outputType, results.ResourceConfig, outputWriter)
+	}
+
 	// add verbose flag to the scan summary
 	results.Violations.ViolationStore.Summary.ShowViolationDetails = s.verbose
 
@@ -235,10 +242,5 @@ func (s ScanOptions) writeResults(results runtime.Output) error {
 		results.Violations.ViolationStore.PassedRules = nil
 	}
 
-	outputWriter := NewOutputWriter(s.UseColors)
-
-	if s.configOnly {
-		return writer.Write(s.outputType, results.ResourceConfig, outputWriter)
-	}
 	return writer.Write(s.outputType, results.Violations, outputWriter)
 }
